SoccerPedia/requirements.txt
flask
requests
django
djangorestframework
drf_yasg
Pillow
boto3
pycountry
openai
psycopg2-binary
django-storages
asyncio
aiohttp
django-cors-headers
whitenoise
channels
djangorestframework-simplejwt
django-allauth